Relocate PubSub to the Server hierarchy
The new pub sub implementation applies only to instances of objects,
which is great for a server (of which you’d only expect to have one),
but pretty pointless for a client (which one might wish to pool, or
reinstantiate, or whatnot). Move it in the class and folder hierarchy
to reflect this.