Switch to Patron for HTTP connections
HTTPClient is nice, but the way it maintains persistent connections
that's difficult to work with. In particular, it offers no way to limit
connection count, simply creating a new one when needed. Switch to
Patron simply because it maintains persistence for a single instance,
allowing us to pool connections in a controlled manner and share them
amongst worker threads.