Change canonical service route generation
`#underscore` is an ActiveSupport method, and we don't need none o'
that. Just swapping out invalid class name characters gives us
everything we want from a URL anyway.

Also, use a more technically accurate regex for RPC names based on the
Thrift IDL, which specifies identifiers as [a-zA-Z0-9._], or [\w.] in
Ruby parlance.