Switch default server protocol to JSON
The JSON thrift protocol isn't especially useful for actual serving, but
it's great for producing readable content to `curl` and similar tools
(without involving a hex decoder). Switch to it on the server side for
ease of debugging while keeping the client on compact protocol; this
ensures that the defaults are nice for debugging _and_ nice for usage.